cloudpilot-ai-backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
monotonic==1.6
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
packaging==23.0
pandas==2.0.0
pansi==2020.7.3
Pygments==2.15.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
